Last sent code from Kalypso
So, this is the code. Bertolino noticed that the function "dist" actually occupies much more space then the older one, so he's gonna change it, making the codesize from 100% to 79%. This way, the little changes that it needed can be done. 

I would say that this code has some potential. We'll need to compare this to the other code from Coffe and O&P. We could merge the best from both teams and actually make something interesting